mobile/kivy_gps_tracker/docker/requirements.txt
python-dotenv
fastapi
uvicorn
python-multipart
pydantic
boto3
requests
starlette
sqlalchemy
passlib
bcrypt
python-jose[cryptography]
aiosqlite
